refactoring the refactor
