--- usaspending_api/common/validator/tests/unit/test_tinyshield_v2.py
@@ -1,256 +1,0 @@
-import copy
-import pytest
-
-from unittest import mock
-
-from usaspending_api.common.exceptions import UnprocessableEntityException
-from usaspending_api.common.validator.award_filter import AWARD_FILTER
-from usaspending_api.common.validator.helpers import validate_array
-from usaspending_api.common.validator.helpers import validate_boolean
-from usaspending_api.common.validator.helpers import validate_datetime
-from usaspending_api.common.validator.helpers import validate_enum
-from usaspending_api.common.validator.helpers import validate_float
-from usaspending_api.common.validator.helpers import validate_integer
-from usaspending_api.common.validator.helpers import validate_object
-from usaspending_api.common.validator.helpers import validate_text
-from usaspending_api.common.validator.tinyshield_v2 import (validation_function, check_models,
-                                                            block, parse_request, recurse_append)
-
-
-ARRAY_RULE = {'name': 'test', 'type': 'array', 'key': 'filters|test',
-              'array_type': 'integer', 'optional': True, 'array_min': 1, 'array_max': 3, 'value': [1, 2, 3]}
-BOOLEAN_RULE = {'name': 'test', 'type': 'boolean', 'key': 'filters|test', 'optional': True, 'value': True}
-DATETIME_RULE = {'name': 'test', 'type': 'datetime', 'key': 'filters|test', 'optional': True,
-                 'value': '1984-09-16T4:05:00'}
-ENUM_RULE = {'name': 'test', 'type': 'enum', 'key': 'filters|test',
-             'enum_values': ['foo', 'bar'], 'optional': True, 'value': "foo"}
-FLOAT_RULE = {'name': 'test', 'type': 'float', 'key': 'filters|test',
-              'optional': True, 'value': 3.14, 'min': 2, 'max': 4}
-TEXT_RULE = {'name': 'test', 'type': 'text', 'key': 'filters|test',
-             'optional': True, 'value': "hello world", "text_type": "search"}
-INTEGER_RULE = {'name': 'test', 'type': 'integer', 'key': 'filters|test',
-                'optional': True, 'value': 3, 'min': 2, 'max': 4}
-OBJECT_RULE = {'name': 'test', 'type': 'object', 'key': 'filters|test',
-               'object_keys': {
-                    'foo': {'type': 'string', 'optional': False},
-                    'hello': {'type': 'integer', 'optional': False}
-                },
-               'value': {'foo': 'bar', 'hello': 1}}
-
-FILTER_OBJ = {
-        "filters": {
-            "keywords": ["grumpy", "bungle"],
-            "award_type_codes": [
-                "A",
-                "B",
-                "C",
-                "D"
-            ],
-            "time_period": [
-                {
-                    "start_date": "2008-01-01",
-                    "end_date": "2011-01-31"
-                }
-            ],
-            "place_of_performance_scope": "domestic",
-            "place_of_performance_locations": [
-                {
-                    "country": "USA",
-                    "state": "VA",
-                    "county": "059"
-                }
-            ],
-            "agencies": [
-                {
-                    "type": "funding",
-                    "tier": "toptier",
-                    "name": "Office of Pizza"
-                },
-                {
-                    "type": "awarding",
-                    "tier": "subtier",
-                    "name": "Personal Pizza"
-                }
-            ],
-            "recipient_search_text": ["D12345678"],
-            "recipient_scope": "domestic",
-            "recipient_locations": [
-                {
-                    "country": "USA",
-                    "state": "VA",
-                    "county": "059"
-                }
-            ],
-            "recipient_type_names": [
-                "Small Business",
-                "Alaskan Native Owned Business"
-            ],
-            "award_ids": ["1605SS17F00018"],
-            "award_amounts": [
-                  {
-                    "lower_bound": 1000000.00,
-                    "upper_bound": 25000000.00
-                  },
-                  {
-                    "upper_bound": 1000000.00
-                  },
-                  {
-                    "lower_bound": 500000000.00
-                  }
-            ],
-            "program_numbers": ["10.553"],
-            "naics_codes": ["336411"],
-            "psc_codes": ["1510"],
-            "contract_pricing_type_codes": ["SAMPLECODE123"],
-            "set_aside_type_codes": ["SAMPLECODE123"],
-            "extent_competed_type_codes": ["SAMPLECODE123"]
-
-        }
-    }
-
-
-'''
-Because these functions all raise Exceptions on failure, all we need to do to write the unit tests is call the function.
-If an exception is raised, the test will fail
-'''
-
-
-def test_validate_array():
-    validate_array(ARRAY_RULE)
-
-
-def test_validate_boolean():
-    validate_boolean(BOOLEAN_RULE)
-
-
-def test_validate_datetime():
-    validate_datetime(DATETIME_RULE)
-
-
-def test_validate_enum():
-    validate_enum(ENUM_RULE)
-
-
-def test_validate_float():
-    validate_float(FLOAT_RULE)
-
-
-def test_validate_integer():
-    validate_integer(INTEGER_RULE)
-
-
-def test_validate_text():
-    validate_text(TEXT_RULE)
-
-
-def test_validate_object():
-    validate_object(OBJECT_RULE)
-
-
-def setup_view(view, request, *args, **kwargs):
-    """
-    Mimic ``as_view()``, but returns view instance.
-    Use this function to get view instances on which you can run unit tests,
-    by testing specific methods.
-    """
-
-    view.request = request
-    view.args = args
-    view.kwargs = kwargs
-    return view
-
-
-def test_check_models():
-    '''We want this test to fail if either AWARD_FILTERS has an invalid model,
-    OR if the logic of the check_models function has been corrupted.
-    '''
-    check_models(copy.deepcopy(AWARD_FILTER))
-
-
-def test_recurse_append():
-    mydict = {}
-    struct = ['level1', 'level2']
-    data = "foobar"
-    recurse_append(struct, mydict, data)
-    assert mydict == {"level1": {"level2": "foobar"}}
-
-
-def test_parse_request():
-    request = copy.deepcopy(FILTER_OBJ)
-    rules = parse_request(request, check_models(copy.deepcopy(AWARD_FILTER)))
-    assert all("value" in item for item in rules)
-
-
-def test_block():
-    request = copy.deepcopy(FILTER_OBJ)
-    request["something"] = "garbage"
-    data = block(request, check_models(copy.deepcopy(AWARD_FILTER)))
-    assert data == FILTER_OBJ
-
-
-def test_validation_function():
-    data = copy.deepcopy(FILTER_OBJ)
-    data["something"] = "garbage"
-    request = mock.Mock(spec=['data'])
-    request.data = data
-    request = validation_function(request, check_models(copy.deepcopy(AWARD_FILTER)))
-    assert request.data == FILTER_OBJ
-
-
-# Test the "any" rule.
-def test_any_rule():
-    models = [{
-        'name': 'value',
-        'key': 'value',
-        'type': 'any',
-        'models': [
-            {'type': 'integer'},
-            {'type': 'text', 'text_type': 'search'}
-        ]
-    }]
-
-    # Test integer and random other key.
-    ts = block({'value': 1, 'another_value': 2}, check_models(models))
-    assert ts['value'] == 1
-    assert ts.get('another_value') is None
-
-    # Test integer masquerading as a string.
-    ts = block({'value': '1'}, check_models(models))
-    assert ts['value'] == 1
-
-    # Test string.
-    ts = block({'value': 'XYZ'}, check_models(models))
-    assert ts['value'] == 'XYZ'
-
-    # Test list (which should blow up).
-    with pytest.raises(UnprocessableEntityException):
-        block({'value': ['XYZ']}, check_models(models))
-
-    # Test with optional 'value' missing.
-    ts = block({'another_value': 2}, check_models(models))
-    assert ts.get('value') is None
-    assert ts.get('another_value') is None
-
-    # Make 'value' required then run the same tests as above.
-    models[0]['optional'] = False
-
-    # Test integer and random other key.
-    ts = block({'value': 1, 'another_value': 2}, check_models(models))
-    assert ts['value'] == 1
-    assert ts.get('another_value') is None
-
-    # Test integer masquerading as a string.
-    ts = block({'value': '1'}, check_models(models))
-    assert ts['value'] == 1
-
-    # Test string.
-    ts = block({'value': 'XYZ'}, check_models(models))
-    assert ts['value'] == 'XYZ'
-
-    # Test list (which should blow up).
-    with pytest.raises(UnprocessableEntityException):
-        block({'value': ['XYZ']}, check_models(models))
-
-    # Test with required 'value' missing.
-    with pytest.raises(UnprocessableEntityException):
-        block({'another_value': 2}, check_models(models))
