+++ usaspending_api/settings.py
@@ -224,6 +224,7 @@ STATIC_URL = '/static/'
 STATIC_ROOT = os.path.join(BASE_DIR, 'usaspending_api/static/')
 STATICFILES_DIRS = (os.path.join(BASE_DIR, 'usaspending_api/static_doc_files'),)
 
+'''
 LOGGING = {
     'version': 1,
     'disable_existing_loggers': False,
@@ -272,6 +273,23 @@ LOGGING = {
             'handlers': ['console', 'console_file'],
             'level': 'INFO',
             'propagate': False,
+        }
+    },
+}
+'''
+LOGGING = {
+    'version': 1,
+    'handlers': {
+        'console':{
+            'level':'DEBUG',
+            'class':'logging.StreamHandler',
+        },
+    },
+    'loggers': {
+        'django.request': {
+            'handlers':['console'],
+            'propagate': True,
+            'level':'DEBUG',
         }
     },
 }
